Refactored: All Datadog::OpenTracing names to OpenTracer.
