+++ Rakefile
@@ -15,11 +15,11 @@ namespace :spec do
 
   RSpec::Core::RakeTask.new(:main) do |t|
     t.pattern = 'spec/**/*_spec.rb'
-    t.exclude_pattern = 'spec/**/{contrib,benchmark,redis,opentracing}/**/*_spec.rb'
+    t.exclude_pattern = 'spec/**/{contrib,benchmark,redis,opentracer}/**/*_spec.rb'
   end
 
-  RSpec::Core::RakeTask.new(:opentracing) do |t|
-    t.pattern = 'spec/ddtrace/opentracing/**/*_spec.rb'
+  RSpec::Core::RakeTask.new(:opentracer) do |t|
+    t.pattern = 'spec/ddtrace/opentracer/**/*_spec.rb'
   end
 
   RSpec::Core::RakeTask.new(:rails) do |t|
@@ -244,7 +244,7 @@ task :ci do
     # Main library
     sh 'rake test:main'
     sh 'rake spec:main'
-    sh 'rake spec:opentracing'
+    sh 'rake spec:opentracer'
 
     if RUBY_PLATFORM != 'java'
       # Contrib minitests
@@ -283,7 +283,7 @@ task :ci do
     # Main library
     sh 'rake test:main'
     sh 'rake spec:main'
-    sh 'rake spec:opentracing'
+    sh 'rake spec:opentracer'
 
     if RUBY_PLATFORM != 'java'
       # Contrib minitests
@@ -328,7 +328,7 @@ task :ci do
     # Main library
     sh 'rake test:main'
     sh 'rake spec:main'
-    sh 'rake spec:opentracing'
+    sh 'rake spec:opentracer'
 
     if RUBY_PLATFORM != 'java'
       # Contrib minitests
@@ -382,7 +382,7 @@ task :ci do
     # Main library
     sh 'rake test:main'
     sh 'rake spec:main'
-    sh 'rake spec:opentracing'
+    sh 'rake spec:opentracer'
 
     if RUBY_PLATFORM != 'java'
       # Contrib minitests
@@ -436,7 +436,7 @@ task :ci do
     # Main library
     sh 'rake test:main'
     sh 'rake spec:main'
-    sh 'rake spec:opentracing'
+    sh 'rake spec:opentracer'
 
     if RUBY_PLATFORM != 'java'
       # Contrib minitests
