+++ lib/ddtrace/opentracer/span_context.rb
@@ -1,5 +1,5 @@
 module Datadog
-  module OpenTracing
+  module OpenTracer
     class SpanContext < ::OpenTracing::SpanContext
     end
   end
