+++ spec/ddtrace/opentracer/helper.rb
@@ -1,0 +1,5 @@
+RSpec.shared_context 'OpenTracing helpers' do
+  before(:each) do
+    skip 'OpenTracing not supported' unless Datadog::OpenTracer.supported?
+  end
+end
