+++ spec/ddtrace/opentracer/scope_manager_spec.rb
@@ -1,17 +1,17 @@
 require 'spec_helper'
 
-require 'ddtrace/opentracing'
-require 'ddtrace/opentracing/helper'
+require 'ddtrace/opentracer'
+require 'ddtrace/opentracer/helper'
 
-if Datadog::OpenTracing.supported?
-  RSpec.describe Datadog::OpenTracing::ScopeManager do
+if Datadog::OpenTracer.supported?
+  RSpec.describe Datadog::OpenTracer::ScopeManager do
     include_context 'OpenTracing helpers'
 
     subject(:scope_manager) { described_class.new }
 
     describe '#activate' do
       subject(:activate) { scope_manager.activate(span, finish_on_close: finish_on_close) }
-      let(:span) { instance_double(Datadog::OpenTracing::Span) }
+      let(:span) { instance_double(Datadog::OpenTracer::Span) }
       let(:finish_on_close) { true }
       it { is_expected.to be(OpenTracing::Scope::NOOP_INSTANCE) }
     end
