--- spec/ddtrace/opentracing/helper.rb
@@ -1,5 +1,0 @@
-RSpec.shared_context 'OpenTracing helpers' do
-  before(:each) do
-    skip 'OpenTracing not supported' unless Datadog::OpenTracing.supported?
-  end
-end
